chore: Refactor frontend utilities files by making a js/utils/ folder which has scope-specific utils.js files such as style-utils.js
