+++ static/js/settings.js
@@ -16,10 +16,15 @@ import {
   saveAppSettingsToServer,
   getTheme,
 } from "./settingsState.js";
-import { formatDistance } from "./utils.js";
+
+import { formatDistance } from "./utils/format-utils.js";
+
 import { logout, deleteAccount } from "./auth/auth.js";
+
 import { applyTheme } from "./ui.js";
+
 import { hasActiveRouteStatsPanel } from "./routes/routeState.js";
+
 import { updateSavedRouteCards } from "./routes/savedRoutesDashboard.js";
 
 /** @type {(() => void) | null} */
